remove store & add context
