+++ src/context/TaskContext.js
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+
+const TaskContext = createContext({
+    tasks: []
+});
+
+export default TaskContext;
